[clients] Warnings

+++ src/libraries/frontend/CommandLineTools.cpp
@@ -16,16 +16,16 @@ namespace frontend
     // Name: SimulationPort
     // Created: SBO 2008-02-20
     // -----------------------------------------------------------------------------
-    unsigned short SimulationPort( unsigned exerciseNumber )
+    unsigned short SimulationPort( unsigned int exerciseNumber )
     {
-        return 10000 + ( exerciseNumber - 1 ) * 2;
+        return static_cast< unsigned short >( 10000 + ( exerciseNumber - 1 ) * 2 );
     }
 
     // -----------------------------------------------------------------------------
     // Name: DispatcherPort
     // Created: SBO 2008-02-20
     // -----------------------------------------------------------------------------
-    unsigned short DispatcherPort( unsigned exerciseNumber )
+    unsigned short DispatcherPort( unsigned int exerciseNumber )
     {
         return SimulationPort( exerciseNumber ) + 1;
     }
